requirements.txt
discord.py
python-dotenv
requests
aiohttp
Flask
asyncpg
apscheduler
pytz